rag/tech-docs-explorer/pyproject.toml
[project]
name = "tech-docs-explorer"
version = "0.1.0"
description = "Educational RAG application for exploring technical documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core>=0.14.12",
    "llama-index-llms-openai>=0.6.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-readers-web>=0.5.6",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-chroma>=0.5.5",
    "chromadb>=1.4.1",

    "pypdf>=6.6.0",
    "streamlit>=1.30.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["config", "core", "llm", "ui", "utils"]
